Refactor user to mock user
